src/requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.7.1
asgiref==3.8.1
asyncpg==0.25.0
base58==2.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coincurve==21.0.0
colorama==0.4.6
cytoolz==1.0.1
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
fastapi==0.78.0
fastapi-pagination==0.9.0
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
parsimonious==0.10.0
psycopg2-binary==2.9.6
pycryptodome==3.22.0
pydantic==1.10.21
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==1.4.36
starlette==0.19.1
toolz==1.0.0
tronpy==0.5.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==15.0.1
